% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20008.R
\name{BS}
\alias{BS}
\title{Bart Simpson density function}
\usage{
BS(x)
}
\arguments{
\item{x}{the chosen interval}
}
\value{
Bart Simpson function\code{n}
}
\description{
Bart Simpson density function f(x)=1/2 phi(x;0,1)+1/10 sum_{j=0}^4 phi(x;j/2-1,1/10)
}
\examples{
\dontrun{
x<-seq(-1,1,0.01)
plot(x,BS(x))
}
}
