% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20008.R
\name{HBS}
\alias{HBS}
\title{Generate samples from Bart Simpson density and draw histogram}
\usage{
HBS(n, h)
}
\arguments{
\item{n}{number of desired random sample points}

\item{h}{value chosen for bandwidth}
}
\value{
Bart Simpson histogram\code{n}
}
\description{
Get random points generated from Bart Simpson density function and draw histogram accordingly.
}
