% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20008.R
\name{NBS}
\alias{NBS}
\title{Use naive density estimator to estimate BS density,}
\usage{
NBS(x, y, h)
}
\arguments{
\item{x}{estimate points}

\item{y}{random samples}

\item{h}{bandwidth}
}
\value{
estimated density at x \code{n}
}
\description{
Use naive density estimator to estimate density,
}
